backend/requirements.txt
Flask==3.0.0
Flask-CORS==4.0.0
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.5
python-dotenv==1.0.0
marshmallow==3.20.1
flask-marshmallow==0.15.0
marshmallow-sqlalchemy==0.29.0
psycopg2-binary==2.9.9
gunicorn==21.2.0
deepgram-sdk==3.2.7
websockets==12.0
asyncio==3.4.3
